+++ codex/librarian/importer/query_fks.py
@@ -48,7 +48,7 @@ _CLASS_QUERY_FIELDS_MAP = {
     Imprint: ("publisher__name", "name"),
     Series: ("publisher__name", "imprint__name", "name"),
     Volume: ("publisher__name", "imprint__name", "series__name", "name"),
-    Identifier: ("identifier_type__name", "nss", "url"),
+    Identifier: ("identifier_type__name", "nss"),
 }
 _DEFAULT_QUERY_FIELDS = ("name",)
 
@@ -75,19 +75,14 @@ class QueryForeignKeysMixin(QueuedThread):
         """Query existing metatata tables."""
         fields = _CLASS_QUERY_FIELDS_MAP.get(fk_cls, _DEFAULT_QUERY_FIELDS)
         flat = len(fields) == 1 and fk_cls != Publisher
-        qs = (
-            fk_cls.objects.filter(fk_filter)
-            .order_by("pk")
-            .values_list(*fields, flat=flat)
-        )
+        qs = fk_cls.objects.filter(fk_filter).values_list(*fields, flat=flat)
         return frozenset(qs)
 
-    def _query_create_metadata(  # noqa: PLR0913
+    def _query_create_metadata(
         self,
         fk_cls,
         create_mds,
         q_obj: Q,
-        q_and_chunk_prefix,
         status,
     ):
         """Get create metadata by comparing proposed meatada to existing rows."""
@@ -101,8 +96,6 @@ class QueryForeignKeysMixin(QueuedThread):
             # django.db.utils.OperationalError: Expression tree is too large (maximum depth 1000)
             children_chunk = q_obj.children[offset : offset + FILTER_BATCH_SIZE]
             filter_chunk = Q(*children_chunk, _connector=Q.OR)
-            if q_and_chunk_prefix:
-                filter_chunk = q_and_chunk_prefix & filter_chunk
 
             existing_mds = self._query_existing_mds(fk_cls, filter_chunk)
             create_mds.difference_update(existing_mds)
@@ -153,7 +146,7 @@ class QueryForeignKeysMixin(QueuedThread):
 
         candidate_groups = set(groups.keys())
         count = self._query_create_metadata(
-            group_cls, candidate_groups, all_filters, None, status
+            group_cls, candidate_groups, all_filters, status
         )
         create_group_set.update(candidate_groups)
         return count
@@ -316,43 +309,39 @@ class QueryForeignKeysMixin(QueuedThread):
     #########
     @staticmethod
     def _query_missing_identifiers_model_obj(
-        key_rel, value_rel_map, item, possible_create_objs
+        value_rel_map, key, values, possible_create_objs, url_restore_map
     ):
         """Update all_filters and possible_create_objs for identifiers."""
         # identifiers is unique at this time for having more than one value as a dict.
-        key, values = item
-        values_list = []
-        for val in values:
-            values_list.append(dict(val))
-
         value_filter = Q()
-        for values_dict in values_list:
+        # value filter
+        for value in values:
+            values_dict = dict(value)
+            create_obj = (key, values_dict.get("nss", ""))
+            url_restore_map[create_obj] = values_dict.get("url", "")
+            possible_create_objs.add(create_obj)
             value_sub_filter = Q()
             for value_key, obj_rel in value_rel_map.items():
-                value = values_dict.get(value_key)
-                value_sub_filter &= Q(**{f"{obj_rel}": value})
+                values_value = values_dict.get(value_key)
+                value_sub_filter &= Q(**{f"{obj_rel}": values_value})
             if value_sub_filter:
                 value_filter |= value_sub_filter
 
-        if not value_filter:
-            return value_filter, None
-
-        obj_filter = Q(**{key_rel: key})
-
-        for value in values:
-            possible_create_objs.add((key, value))
-        return value_filter, obj_filter
+        return value_filter
 
     @staticmethod
-    def _query_missing_dict_model_obj(key_rel, value_rel, item, possible_create_objs):
+    def _query_missing_dict_model_obj(
+        value_rel, key, values, possible_create_objs, _url_restore_map
+    ):
         """Update all_filters and possible_create_objs for this obj."""
-        key, values = item
+        for value in values:
+            possible_create_objs.add((key, value))
 
         filter_isnull = None in values
         filter_values = values - {None}
 
         if not filter_values and not filter_isnull:
-            return Q(), None
+            return Q()
 
         value_filter = (
             Q(**{f"{value_rel}__in": filter_values}) if filter_values else Q()
@@ -360,11 +349,7 @@ class QueryForeignKeysMixin(QueuedThread):
         if filter_isnull:
             value_filter = value_filter | Q(**{f"{value_rel}__isnull": True})
 
-        obj_filter = Q(**{key_rel: key})
-
-        for value in values:
-            possible_create_objs.add((key, value))
-        return value_filter, obj_filter
+        return value_filter
 
     @staticmethod
     def _add_value_filter_map(value_filter, filter_and_prefix, query_filter_map):
@@ -392,22 +377,38 @@ class QueryForeignKeysMixin(QueuedThread):
         else:
             rel_map = DICT_MODEL_REL_MAP
             query_missing_method = self._query_missing_dict_model_obj
+            url_restore_map = {}
 
         model, key_rel, value_rel = rel_map[field_name]
 
-        for item in possible_objs.items():
-            value_filter, filter_and_prefix = query_missing_method(
-                key_rel, value_rel, item, possible_create_objs
+        # this lets an identifier get recreated.
+        # fix that first
+        # Then try to do it with one query and one create.
+
+        url_restore_map = {}
+        for key, values in possible_objs.items():
+            filter_and_prefix = Q(**{key_rel: key})
+            value_filter = query_missing_method(
+                value_rel, key, values, possible_create_objs, url_restore_map
             )
             self._add_value_filter_map(
                 value_filter, filter_and_prefix, query_filter_map
             )
 
         # get the obj metadata
+        combined_q_obj = Q()
         for filter_and_prefix, value_filter in query_filter_map.items():
-            count += self._query_create_metadata(
-                model, possible_create_objs, value_filter, filter_and_prefix, status
-            )
+            combined_q_obj |= filter_and_prefix & value_filter
+
+        count += self._query_create_metadata(
+            model, possible_create_objs, combined_q_obj, status
+        )
+        if field_name == "identifiers":
+            restored_create_objs = []
+            for create_obj in possible_create_objs:
+                url = url_restore_map.get(create_obj)
+                restored_create_objs.append((*create_obj, url))
+            possible_create_objs = restored_create_objs
 
         create_objs.update(possible_create_objs)
         count = len(create_objs)