+++ NEWS.md
@@ -21,6 +21,7 @@ border-radius: 128px;
   - OPDS http basic authorization fixed.
   - Groups with the same name case insensitively collapsed in the browser.
   - Order By respects browser show groups settings.
+  - Fixed re-import of urls and identifiers.
   - Fixed cleanup of some foreign keys when no longer used.
 
 ## v1.5.17